backend/requirements.txt
# Python dependencies for Rahnuma ML service
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
scipy==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.4
httpx==0.27.0
